♻️ Add examaples/namespace.go for Namespace
