+++ example/main.go
@@ -1,32 +1,18 @@
 package main
 
 import (
-	"fmt"
-	"io"
-	"log"
-	"log/slog"
-	"net/http"
 	"os"
-
-	"github.com/akm/go-requestid"
-	"github.com/akm/slogw"
 )
 
 func main() {
-	helloHandler := func(w http.ResponseWriter, req *http.Request) {
-		ctx := req.Context()
-		slog.InfoContext(ctx, "Start")
-		defer slog.InfoContext(ctx, "End")
-		io.WriteString(w, "Hello, world!\n") // nolint: errcheck
+	if len(os.Args) < 2 {
+		basic()
+		return
 	}
-
-	slog.SetDefault(slogw.New(slog.NewTextHandler(os.Stdout, nil)))
-
-	http.Handle("/hello", requestid.Wrap(http.HandlerFunc(helloHandler)))
-	fmt.Println("Server started at :8080")
-	log.Fatal(http.ListenAndServe(":8080", nil))
-}
-
-func init() {
-	slogw.Register(requestid.SlogwPrepareFunc("requestid"))
+	switch os.Args[1] {
+	case "basic":
+		basic()
+	case "namespace":
+		namespace()
+	}
 }
